+++ Finalaplication/Controllers/VolunteerController.cs
@@ -7,6 +7,9 @@ using Finalaplication.App_Start;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Http;
 using System.Text;
+using System;
+using System.Threading.Tasks;
+using ReflectionIT.Mvc.Paging;
 
 namespace Finalaplication.Controllers
 {
@@ -57,7 +60,7 @@ namespace Finalaplication.Controllers
                             Volunteer.ContactInformation.MailAdress,
                             Volunteer.ContactInformation.PhoneNumber)
 
-                            
+
                             }
                              ).ToList();
 
@@ -76,20 +79,94 @@ namespace Finalaplication.Controllers
 
         }
 
-
-
-        public ActionResult Index(string searching)
+        public ActionResult Index(string sortOrder, string searching, bool Active, bool HasCar, bool HasContract, DateTime lowerdate, DateTime upperdate, int page)
         {
+            ViewBag.searching = searching;
+            ViewBag.active = Active;
+            ViewBag.Page = page;
+            ViewBag.Upperdate = upperdate;
+            ViewBag.Lowerdate = lowerdate;
+            ViewBag.hascar = HasCar;
+            ViewBag.hascontract = HasContract;
+
+            ViewBag.NameSortParm = String.IsNullOrEmpty(sortOrder) ? "name_desc" : "";
+            ViewBag.DateSortParm = sortOrder == "Date" ? "date_desc" : "Date";
+            ViewBag.LastnameSort = sortOrder == "Lastname" ? "Lastname_desc" : "Lastname";
+            ViewBag.HourCountSort = sortOrder == "Hourcount" ? "Hourcount_desc" : "Hourcount";
+            ViewBag.Gendersort = sortOrder == "Gender" ? "Gender_desc" : "Gender";
+            ViewBag.Activesort = sortOrder == "Active" ? "Active_desc" : "Active";
+
             List<Volunteer> volunteers = vollunteercollection.AsQueryable().ToList();
+            DateTime d1 = new DateTime(0003, 1, 1);
+            if (upperdate > d1)
+            {
+                volunteers = volunteers.Where(x => x.Birthdate <= upperdate).ToList();
+            }
             if (searching != null)
             {
-                return View(volunteers.Where(x => x.Firstname.Contains(searching) || x.Lastname.Contains(searching)).ToList());
+                volunteers = volunteers.Where(x => x.Firstname.Contains(searching) || x.Lastname.Contains(searching)).ToList();
             }
-            else
+            if (Active == true)
             {
-                return View(volunteers);
+                volunteers = volunteers.Where(x => x.InActivity == true).ToList();
             }
+            if (lowerdate != null)
+            {
+                volunteers = volunteers.Where(x => x.Birthdate > lowerdate).ToList();
+            }
+            if (HasCar == true)
+            {
+                volunteers = volunteers.Where(x => x.Additionalinfo.HasCar == true).ToList();
+            }
+            if (HasContract == true)
+            {
+                volunteers = volunteers.Where(x => x.Contract.HasContract == true).ToList();
+            }
+            switch (sortOrder)
+            {
+                case "Gender":
+                    volunteers = volunteers.OrderBy(s => s.Gender).ToList();
+                    break;
+                case "Gender_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.Gender).ToList();
+                    break;
+                case "Lastname":
+                    volunteers = volunteers.OrderBy(s => s.Lastname).ToList();
+                    break;
+                case "Lastname_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.Lastname).ToList();
+                    break;
+                case "Hourcount":
+                    volunteers = volunteers.OrderBy(s => s.HourCount).ToList();
+                    break;
+                case "Hourcount_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.HourCount).ToList();
+                    break;
+                case "Active":
+                    volunteers = volunteers.OrderBy(s => s.InActivity).ToList();
+                    break;
+                case "Active_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.InActivity).ToList();
+                    break;
+                case "name_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.Firstname).ToList();
+                    break;
+                case "Date":
+                    volunteers = volunteers.OrderBy(s => s.Birthdate).ToList();
+                    break;
+                case "date_desc":
+                    volunteers = volunteers.OrderByDescending(s => s.Birthdate).ToList();
+                    break;
+                default:
+                    volunteers = volunteers.OrderBy(s => s.Firstname).ToList();
+                    break;
+            }
+            ViewBag.counter = volunteers.Count();
+            volunteers = volunteers.AsQueryable().Skip((page - 1) * 5).ToList();
+            volunteers = volunteers.AsQueryable().Take(5).ToList();
+            return View(volunteers);
         }
+
 
         public ActionResult Birthday()
         {
@@ -139,7 +216,7 @@ namespace Finalaplication.Controllers
                 }
                 else
                 {
-                return View();
+                    return View();
                 }
             }
             catch
@@ -154,6 +231,11 @@ namespace Finalaplication.Controllers
             var volunteerId = new ObjectId(id);
             var volunteer = vollunteercollection.AsQueryable<Volunteer>().SingleOrDefault(x => x.VolunteerID == volunteerId);
             return View(volunteer);
+        }
+
+        public ActionResult Volunteerwarning()
+        {
+            return View();
         }
 
         // POST: Volunteer/Edit/5
@@ -213,13 +295,26 @@ namespace Finalaplication.Controllers
 
         // POST: Volunteer/Delete/5
         [HttpPost]
-        public ActionResult Delete(string id, IFormCollection collection)
+        public ActionResult Delete(string id, IFormCollection collection, Volunteer volunteer, bool Inactive)
         {
             try
             {
-                vollunteercollection.DeleteOne(Builders<Volunteer>.Filter.Eq("_id", ObjectId.Parse(id)));
+                if (Inactive == false)
+                {
 
-                return RedirectToAction("Index");
+                    vollunteercollection.DeleteOne(Builders<Volunteer>.Filter.Eq("_id", ObjectId.Parse(id)));
+
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    var filter = Builders<Volunteer>.Filter.Eq("_id", ObjectId.Parse(id));
+                    var update = Builders<Volunteer>.Update
+                        .Set("InActivity", volunteer.InActivity);
+                    var result = vollunteercollection.UpdateOne(filter, update);
+                    return RedirectToAction("Index");
+                }
+
             }
             catch
             {
